api/requirements.txt
django
djangorestframework
django-cors-headers
google-generativeai
python-dotenv
langchain
PyPDF2
faiss-cpu
langchain_google_genai
SpeechRecognition
pyttsx3